daemon false

+++ utils/helper.go
@@ -4,6 +4,6 @@ import (
     "time"
 )
 
-func UnixTime() int {
-    return time.Now().Second()
+func UnixTime() int64 {
+    return int64(time.Now().Second())
 }
